refactor actions
