+++ client/src/redux/_actions/index.js
@@ -1,2 +1,3 @@
 export * from './chat.actions.js';
 export * from './file.actions.js';
+export * from './streams.actions.js';
